Disable mcontext patching
I used to have a patch against tccrun.c that was supposed to make
it possible for tcc to compile on NetBSD and Cygwin. The patch for
NetBSD made it into upstream, and the regular expression I had used
for finding the patch spot no longer applies to the source file
(because the source file was modified upstream). It hasn't been
applying anyway, so I'm leaving it in there in case I want to use
it in the future, but I'm explicitly disabling it until that time
arrives.